Split Risc0Guest on native/zkvm implementations
