+++ crates/sovereign-sdk/adapters/risc0/src/guest/native.rs
@@ -1,36 +1,15 @@
 //! This module implements the `ZkvmGuest` trait for the RISC0 VM.
-#[cfg(not(target_os = "zkvm"))]
 use std::ops::DerefMut;
 
-#[cfg(target_os = "zkvm")]
-use risc0_zkvm::guest::env;
-#[cfg(not(target_os = "zkvm"))]
 use risc0_zkvm::serde::{Deserializer, WordRead};
-use serde::de::DeserializeOwned;
-use serde::Serialize;
-use sov_rollup_interface::zk::{Zkvm, ZkvmGuest};
+use sov_rollup_interface::zk::ZkvmGuest;
 
-use crate::Risc0MethodId;
-
-#[cfg(target_os = "zkvm")]
-impl ZkvmGuest for Risc0Guest {
-    fn read_from_host<T: serde::de::DeserializeOwned>(&self) -> T {
-        env::read()
-    }
-
-    fn commit<T: serde::Serialize>(&self, item: &T) {
-        env::commit(item);
-    }
-}
-
-#[cfg(not(target_os = "zkvm"))]
 #[derive(Default)]
 struct Hints {
     values: Vec<u32>,
     position: usize,
 }
 
-#[cfg(not(target_os = "zkvm"))]
 impl Hints {
     pub fn with_hints(hints: Vec<u32>) -> Self {
         Hints {
@@ -40,7 +19,6 @@ impl Hints {
     }
 }
 
-#[cfg(not(target_os = "zkvm"))]
 impl WordRead for Hints {
     fn read_words(&mut self, words: &mut [u32]) -> risc0_zkvm::serde::Result<()> {
         if let Some(slice) = self.values.get(self.position..self.position + words.len()) {
@@ -66,15 +44,11 @@ impl WordRead for Hints {
     }
 }
 
-/// A guest for the RISC0 VM. When running in the Risc0 environment, this struct
-/// implements the `ZkvmGuest` trait in terms of Risc0's env::read and env::commit functions.
-/// When running in any other environment, the struct uses interior mutability to emulate
-/// the same functionality.
+/// A guest for the RISC0 VM. Implements the `ZkvmGuest` trait
+/// using interior mutability to test the functionality.
 #[derive(Default)]
 pub struct Risc0Guest {
-    #[cfg(not(target_os = "zkvm"))]
     hints: std::sync::Mutex<Hints>,
-    #[cfg(not(target_os = "zkvm"))]
     commits: std::sync::Mutex<Vec<u32>>,
 }
 
@@ -85,9 +59,6 @@ impl Risc0Guest {
     }
 
     /// Constructs a new Risc0 Guest with the provided hints.
-    ///
-    /// This function is only available outside of Risc0's environment.
-    #[cfg(not(target_os = "zkvm"))]
     pub fn with_hints(hints: Vec<u32>) -> Self {
         Self {
             hints: std::sync::Mutex::new(Hints::with_hints(hints)),
@@ -96,7 +67,6 @@ impl Risc0Guest {
     }
 }
 
-#[cfg(not(target_os = "zkvm"))]
 impl ZkvmGuest for Risc0Guest {
     fn read_from_host<T: serde::de::DeserializeOwned>(&self) -> T {
         let mut hints = self.hints.lock().unwrap();
@@ -108,29 +78,5 @@ impl ZkvmGuest for Risc0Guest {
         self.commits.lock().unwrap().extend_from_slice(
             &risc0_zkvm::serde::to_vec(item).expect("Serialization to vec is infallible"),
         );
-    }
-}
-
-impl Zkvm for Risc0Guest {
-    type CodeCommitment = Risc0MethodId;
-
-    type Error = anyhow::Error;
-
-    fn verify<'a>(
-        _serialized_proof: &'a [u8],
-        _code_commitment: &Self::CodeCommitment,
-    ) -> Result<&'a [u8], Self::Error> {
-        // Implement this method once risc0 supports recursion: issue #633
-        todo!("Implement once risc0 supports recursion: https://github.com/Sovereign-Labs/sovereign-sdk/issues/633")
-    }
-
-    fn verify_and_extract_output<
-        Da: sov_rollup_interface::da::DaSpec,
-        Root: Serialize + DeserializeOwned,
-    >(
-        _serialized_proof: &[u8],
-        _code_commitment: &Self::CodeCommitment,
-    ) -> Result<sov_rollup_interface::zk::StateTransition<Da, Root>, Self::Error> {
-        todo!()
     }
 }
