+++ crates/sovereign-sdk/adapters/risc0/src/guest/zkvm.rs
@@ -1,0 +1,25 @@
+//! This module implements the `ZkvmGuest` trait for the RISC0 VM.
+use risc0_zkvm::guest::env;
+use sov_rollup_interface::zk::ZkvmGuest;
+
+/// A guest for the RISC0 VM. Implements the `ZkvmGuest` trait
+///  in terms of Risc0's env::read and env::commit functions.
+#[derive(Default)]
+pub struct Risc0Guest {}
+
+impl Risc0Guest {
+    /// Constructs a new Risc0 Guest
+    pub fn new() -> Self {
+        Self::default()
+    }
+}
+
+impl ZkvmGuest for Risc0Guest {
+    fn read_from_host<T: serde::de::DeserializeOwned>(&self) -> T {
+        env::read()
+    }
+
+    fn commit<T: serde::Serialize>(&self, item: &T) {
+        env::commit(item);
+    }
+}
